requirements.txt
flask>=2.0.0
flask-cors>=3.0.0
requests>=2.25.0
python-dotenv>=1.0.0
rich>=12.0.0
aiohttp>=3.8.0
supabase>=2.6.0
boto3>=1.20.0
pillow>=10.4.0
gunicorn>=20.1.0
sqlalchemy>=1.4.0
httpx>=0.24.0
web3>=6.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.28.0
pyyaml>=6.0
pydantic>=2.0.0

# Render deployment fix - v2.0
# Added missing yaml and pydantic dependencies
